+++ include/samurai/mr/adapt.hpp
@@ -298,101 +298,6 @@ namespace samurai
             update_tag_subdomains(level, m_tag);
         }
 
-        // FIXME: this graduation doesn't make the same that the lines below:
-        // why? graduation(m_tag,
-        // stencil_graduation::call(samurai::Dim<dim>{}));
-
-        //// COARSENING GRADUATION
-        // for (std::size_t level = max_level; level > 0; --level)
-        //{
-        //     auto keep_subset = intersection(mesh[mesh_id_t::cells][level], mesh[mesh_id_t::all_cells][level - 1]).on(level - 1);
-        //
-        //    keep_subset.apply_op(maximum(m_tag));
-        //
-        //    int grad_width = static_cast<int>(mesh_t::config::graduation_width);
-        //    auto stencil   = detail::box_dir<dim>();
-        //
-        //    for (int ig = 1; ig <= grad_width; ++ig)
-        //    {
-        //        for (std::size_t is = 0; is < stencil.shape(0); ++is)
-        //        {
-        //            auto s = ig * xt::view(stencil, is);
-        //            auto subset = intersection(mesh[mesh_id_t::cells][level], translate(mesh[mesh_id_t::all_cells][level - 1],
-        //            s)).on(level - 1); subset.apply_op(balance_2to1(m_tag, s));
-        //        }
-        //    }
-        //
-        //    update_tag_periodic(level, m_tag);
-        //    update_tag_subdomains(level, m_tag);
-        //}
-        //
-        //// REFINEMENT GRADUATION
-        // for (std::size_t level = max_level; level > min_level; --level)
-        //{
-        //     auto subset_1 = intersection(mesh[mesh_id_t::cells][level], mesh[mesh_id_t::cells][level]);
-        //
-        //    subset_1.apply_op(extend(m_tag));
-        //    update_tag_periodic(level, m_tag);
-        //    update_tag_subdomains<false>(level, m_tag);
-        //
-        //    int grad_width = static_cast<int>(mesh_t::config::graduation_width);
-        //    auto stencil   = detail::box_dir<dim>();
-        //
-        //    for (int ig = 1; ig <= grad_width; ++ig)
-        //    {
-        //        for (std::size_t is = 0; is < stencil.shape(0); ++is)
-        //        {
-        //            auto s = ig * xt::view(stencil, is);
-        //            auto subset = intersection(translate(mesh[mesh_id_t::cells][level], s), mesh[mesh_id_t::all_cells][level -
-        //            1]).on(level);
-        //
-        //            subset.apply_op(make_graduation(m_tag));
-        //        }
-        //    }
-        //
-        //    update_tag_periodic(level, m_tag);
-        //    update_tag_subdomains<false>(level, m_tag);
-        //}
-        // update_tag_subdomains<false>(min_level, m_tag);
-        //
-        //// Prevents the coarsening of child cells where the parent intersects the boundary.
-        ////
-        ////   outside           |   inside
-        ////   =======           |   ======
-        ////   tag this cell to  |
-        ////   keep to avoid     |
-        ////   coarsening        |
-        ////                     |
-        ////   level l+1   |-----|-----|
-        ////                     |
-        ////   level l     |-----------|
-        ////
-        //// for (std::size_t level = mesh[mesh_id_t::cells].min_level(); level <= mesh[mesh_id_t::cells].max_level(); ++level)
-        //// {
-        ////     auto set = difference(mesh[mesh_id_t::reference][level], mesh.domain()).on(level);
-        ////     set(
-        ////         [&](const auto& i, const auto& index)
-        ////         {
-        ////             m_tag(level, i, index) = static_cast<int>(CellFlag::keep);
-        ////         });
-        //// }
-        //
-        // for (std::size_t level = max_level; level > 0; --level)
-        //{
-        //     auto keep_subset = intersection(mesh[mesh_id_t::cells][level], mesh[mesh_id_t::all_cells][level - 1]).on(level - 1);
-        //
-        //    keep_subset.apply_op(maximum(m_tag));
-        //    update_tag_periodic(level, m_tag);
-        //    // update_tag_subdomains(level, m_tag);
-        //    update_tag_subdomains<false>(level, m_tag);
-        //}
-        // times::timers.stop("mesh adaptation");
-        // update_ghost_mr(other_fields...);
-        // times::timers.start("mesh adaptation");
-        //
-        // keep_only_one_coarse_tag(m_tag);
-        // return update_field_mr(m_tag, m_fields, other_fields...);
-
         for (std::size_t level = max_level; level > 0; --level)
         {
             auto keep_subset = intersection(mesh[mesh_id_t::cells][level], mesh[mesh_id_t::all_cells][level - 1]).on(level - 1);
