+++ include/samurai/utils.hpp
@@ -4,133 +4,15 @@
 #pragma once
 
 #include <algorithm>
-#include <cassert>
 #include <functional>
 #include <tuple>
 #include <type_traits>
-#include <xtensor/xfunction.hpp>
 
 namespace samurai
 {
     template <class value_t, class index_t>
     struct Interval;
 }
-
-// class OrthGivensMatrixVectorProd
-//{
-//   public:
-//
-//     OrthGivensMatrixVectorProd() = default;
-//
-//     OrthGivensMatrixVectorProd(const size_t i, const size_t j)
-//         : m_i(i)
-//         , m_j(j)
-//     {
-//     }
-//
-//     void set(const size_t i, const size_t j)
-//     {
-//         m_i = i;
-//         m_j = j;
-//     }
-//
-//     template <typename T, size_t N>
-//     xt::xtensor_fixed<T, xt::xshape<N>> operator()(const xt::xtensor_fixed<T, xt::xshape<N>>& x) const
-//     {
-//         xt::xtensor_fixed<T, xt::xshape<N>> y(x);
-//         y[m_i] = -x[m_j];
-//         y[m_j] = x[m_i];
-//
-//         return y;
-//     }
-//
-//     template <typename... Ts>
-//     xt::xtensor<typename xt::xfunction<Ts...>::value_type, 1> operator()(const xt::xfunction<Ts...>& x) const
-//     {
-//         xt::xtensor<typename xt::xfunction<Ts...>::value_type, 1> y(x);
-//         y[m_i] = -x[m_j];
-//         y[m_j] = x[m_i];
-//
-//         return y;
-//     }
-//
-//   private:
-//
-//     size_t m_i = 0;
-//     size_t m_j = 1;
-// };
-//
-// template <size_t dim, typename Function>
-// inline Function forall_orth_givens_rotations(std::integral_constant<size_t, dim>, Function&& func)
-//{
-//     static_assert(dim > 1, "there is no givens rotations in 1D");
-//
-//     OrthGivensMatrixVectorProd G;
-//     for (size_t i = 0; i != dim; ++i)
-//     {
-//         for (size_t j = i + 1; j != dim; ++j)
-//         {
-//             G.set(i, j);
-//             func(G);
-//         }
-//     }
-//     return func;
-// }
-//
-// template <typename Function>
-// inline Function forall_orth_givens_rotations(size_t dim, Function&& func)
-//{
-//     assert(dim > 1);
-//
-//     OrthGivensMatrixVectorProd G;
-//     for (size_t i = 0; i != dim; ++i)
-//     {
-//         for (size_t j = i + 1; j != dim; ++j)
-//         {
-//             G.set(i, j);
-//             func(G);
-//         }
-//     }
-//     return func;
-// }
-//
-// template <size_t dim, typename Function>
-// inline Function forall_unit_vectors(std::integral_constant<size_t, dim>, Function&& func)
-//{
-//     using vector = xt::xtensor_fixed<int, xt::xshape<dim>>;
-//
-//     vector v;
-//     for (size_t i = 0; i != dim; ++i)
-//     {
-//         v[i] = 0;
-//     }
-//     for (size_t i = 0; i != dim; ++i)
-//     {
-//         v[i] = 1;
-//         func(v);
-//         v[i] = 0;
-//     }
-//     return func;
-// }
-//
-// template <typename Function>
-// inline Function forall_unit_vectors(size_t dim, Function&& func)
-//{
-//     using vector = xt::xtensor<int, 1>;
-//
-//     vector v(dim);
-//     for (size_t i = 0; i != dim; ++i)
-//     {
-//         v[i] = 0;
-//     }
-//     for (size_t i = 0; i != dim; ++i)
-//     {
-//         v[i] = 1;
-//         func(v);
-//         v[i] = 0;
-//     }
-//     return func;
-// }
 
 template <typename T, typename Compare>
 inline void sort_indexes(const std::vector<T>& v, Compare cmp, std::vector<size_t>& idx)
